+++ apps/web/app/(homepages)/@logo/default.tsx
@@ -1,0 +1,3 @@
+export function Default() {
+  return null;
+}
