+++ apps/web/app/(homepages)/@logo/page.tsx
@@ -1,0 +1,8 @@
+import { FetchLoadArgs } from "~/lib/utils";
+
+interface Props {
+  params: Pick<FetchLoadArgs, "network">;
+}
+export default function HomeLogo({ params }: Props) {
+  return <></>;
+}
