+++ apps/web/app/(homepages)/[network]/page.tsx
@@ -1,0 +1,8 @@
+import { FetchLoadArgs } from "~/lib/utils";
+
+interface Props {
+  params: Pick<FetchLoadArgs, "network">;
+}
+export function NetworkLogo({ params }: Props) {
+  return <>Logo Network</>;
+}
