+++ apps/web/postcss.config.js
@@ -1,8 +1,6 @@
-const config = require("tailwind-config/tailwind.config.js");
-
 module.exports = {
   plugins: {
-    tailwindcss: { config },
+    tailwindcss: {},
     autoprefixer: {},
   },
-}
+};
