+++ apps/web/styles/globals.css
@@ -4,87 +4,22 @@
 
 @layer base {
   :root {
-    --background: 0 0% 100%;
-    --foreground: 0 0% 3.9%;
-
-    --muted: 0 0% 96.1%;
-    --muted-foreground: 0 0% 45.1%;
-
-    --popover: 0 0% 100%;
-    --popover-foreground: 0 0% 3.9%;
-
-    --card: 0 0% 100%;
-    --card-foreground: 0 0% 3.9%;
-
-    --border: 0 0% 89.8%;
-    --input: 0 0% 89.8%;
-
-    --primary: 0 0% 9%;
-    --primary-foreground: 0 0% 98%;
-
-    --secondary: 0 0% 96.1%;
-    --secondary-foreground: 0 0% 9%;
-
-    --accent: 0 0% 96.1%;
-    --accent-foreground: 0 0% 9%;
-
-    --destructive: 0 84.2% 60.2%;
-    --destructive-foreground: 0 0% 98%;
-
-    --ring: 0 0% 63.9%;
-
-    --radius: 0.5rem;
-  }
-
-  .dark {
-    --background: 0 0% 3.9%;
-    --foreground: 0 0% 98%;
-
-    --muted: 0 0% 14.9%;
-    --muted-foreground: 0 0% 63.9%;
-
-    --popover: 0 0% 3.9%;
-    --popover-foreground: 0 0% 98%;
-
-    --card: 0 0% 3.9%;
-    --card-foreground: 0 0% 98%;
-
-    --border: 0 0% 14.9%;
-    --input: 0 0% 14.9%;
-
-    --primary: 0 0% 98%;
-    --primary-foreground: 0 0% 9%;
-
-    --secondary: 0 0% 14.9%;
-    --secondary-foreground: 0 0% 98%;
-
-    --accent: 0 0% 14.9%;
-    --accent-foreground: 0 0% 98%;
-
-    --destructive: 0 62.8% 30.6%;
-    --destructive-foreground: 0 85.7% 97.3%;
-
-    --ring: 0 0% 14.9%;
-  }
-}
-
-@layer base {
-  * {
-    @apply border-border;
-  }
-  body {
-    @apply bg-background text-foreground;
+    --color-foreground: 236 15% 16%;
+    --color-neutral: 224 11% 45%;
+    --color-muted: 222 10% 55%;
   }
 }
 
 @layer utilities {
-    .lifting-shadow {
-        box-shadow: 0px 3px 6px rgba(42, 43, 46, 0.07), 0px 1px 2px rgba(42, 43, 46, 0.04);
-    }
+  .lifting-shadow {
+    box-shadow:
+      0px 3px 6px rgba(42, 43, 46, 0.07),
+      0px 1px 2px rgba(42, 43, 46, 0.04);
+  }
 
-    .radial-bg {
-        background-image: url("../public//images//radial-bg.svg");
-        background-repeat: no-repeat;
-        background-position: center top;
-    }
+  .radial-bg {
+    background-image: url("../public//images//radial-bg.svg");
+    background-repeat: no-repeat;
+    background-position: center top;
+  }
 }
