+++ apps/web/ui/mobile-actions/index.tsx
@@ -5,7 +5,7 @@ import { OptionGroups } from "~/lib/utils";
 import { DialogPanel } from "~/ui/dialog-panel";
 import SvgRightSidebarOff from "~/ui/icons/RightSidebarOff";
 import SvgSearchOff from "~/ui/icons/SearchOff";
-import { Search } from "~/ui/search";
+import { Search } from "~/ui/old-search";
 
 type Props = {
   searchOptions: OptionGroups;
